+++ router/musicRoute.js
@@ -1,6 +1,9 @@
 const express = require('express')
 const route = express.Router()
 const musicController = require('../controller/musicController')
+const authController = require('../controller/authController')
+const cookieParse = require('cookie-parser')
+route.use(cookieParse()) // Requires to read cookie with req.cookies.
 
 route.get('/', musicController.allMusic)
 
@@ -10,9 +13,15 @@ route.post('/addMusic', express.urlencoded({ extended: true }), musicController.
 
 route.get('/searchMusic', musicController.searchMusic)
 
-route.get('/login', (req, res) => res.render('login'))
+route.get('/login', (req, res) => {
+  console.log(req.cookies.key) // See in console.log a data cookie.
+  if (req.cookies.key !== undefined) {
+    return res.render('gerenciarConta')
+  }
+  res.render('login')
+})
 
-route.post('/login', express.urlencoded({ extended: true }), musicController.login)
+route.post('/login', express.urlencoded({ extended: true }), authController.login)
 
 route.get('/deleteMusic', musicController.allDelete)
