+++ template/module/header.ejs
@@ -17,5 +17,5 @@
          <input name="artist" placeholder="Artista" class="form-control mr-2 mt-3" required>
       </form>
       <a href="./"><div class="botao btn btn-primary mr-2 mt-3" >Inicio</div></a>
-      <a href="./admin"><div class="botao btn btn-primary mt-3">Conta</div></a>
+      <a href="./login"><div class="botao btn btn-primary mt-3">Conta</div></a>
    </div>
